gamewindow created

+++ app/src/main/java/Factory/BasicTableFactory.java
@@ -7,7 +7,7 @@ import Pool.Table;
 public class BasicTableFactory implements TableFactory {
 
     @Override
-    public Table create(double x, double y, double friction, String colour) {
+    public Table create(Long x, Long y, double friction, String colour) {
         return new BasicTable(x, y, friction, colour);
     }
 }
